refactor dilation2d relay
