+++ docs/langref/relay_op.rst
@@ -70,7 +70,6 @@ This level enables typical convnet models.
    tvm.relay.nn.conv2d
    tvm.relay.nn.conv2d_transpose
    tvm.relay.nn.dense
-   tvm.relay.nn.dilation2d
    tvm.relay.nn.max_pool2d
    tvm.relay.nn.max_pool3d
    tvm.relay.nn.avg_pool2d
@@ -171,6 +170,7 @@ This level enables additional math and transform operators.
 
    tvm.relay.image.resize
    tvm.relay.image.crop_and_resize
+   tvm.relay.image.dilation2d
    tvm.relay.vision.multibox_prior
    tvm.relay.vision.multibox_transform_loc
    tvm.relay.vision.nms
@@ -250,7 +250,6 @@ Level 2 Definitions
 .. autofunction:: tvm.relay.nn.conv2d
 .. autofunction:: tvm.relay.nn.conv2d_transpose
 .. autofunction:: tvm.relay.nn.dense
-.. autofunction:: tvm.relay.nn.dilation2d
 .. autofunction:: tvm.relay.nn.max_pool2d
 .. autofunction:: tvm.relay.nn.max_pool3d
 .. autofunction:: tvm.relay.nn.avg_pool2d
@@ -339,6 +338,7 @@ Level 5 Definitions
 -------------------
 .. autofunction:: tvm.relay.image.resize
 .. autofunction:: tvm.relay.image.crop_and_resize
+.. autofunction:: tvm.relay.image.dilation2d
 .. autofunction:: tvm.relay.vision.multibox_prior
 .. autofunction:: tvm.relay.vision.multibox_transform_loc
 .. autofunction:: tvm.relay.vision.nms
