+++ include/tvm/relay/attrs/nn.h
@@ -155,43 +155,6 @@ struct Conv2DAttrs : public tvm::AttrsNode<Conv2DAttrs> {
   }
 };
 
-
-/*! \brief Attributes used in dilation operators */
-struct Dilation2DAttrs : public tvm::AttrsNode<Dilation2DAttrs> {
-  Array<IndexExpr> strides;
-  Array<IndexExpr> padding;
-  Array<IndexExpr> dilations;
-  std::string data_layout;
-  std::string kernel_layout;
-  DataType out_dtype;
-
-  TVM_DECLARE_ATTRS(Dilation2DAttrs, "relay.attrs.Dilation2DAttrs") {
-    TVM_ATTR_FIELD(strides).set_default(Array<IndexExpr>({1, 1}))
-        .describe("Specifies the strides of the sliding window. [stride_height, stride_width].");
-    TVM_ATTR_FIELD(padding).set_default(Array<IndexExpr>({0, 0}))
-        .describe("If padding is non-zero, then the input is implicitly zero-padded"
-                  "Padding support both symmetric and asymmetric as"
-                  "one int : same padding used on all sides"
-                  "two int : bottom, right will use same padding as top, left"
-                  "four int : padding width in the order of (top, left, bottom, right)");
-    TVM_ATTR_FIELD(dilations).set_default(Array<IndexExpr>({1, 1}))
-        .describe("Specifies the dilation rate to use. [dilation_height, dilation_width]");
-    TVM_ATTR_FIELD(data_layout).set_default("NCHW")
-        .describe("Dimension ordering of input data. Can be 'NCHW', 'NHWC', etc."
-                  "'N', 'C', 'H', 'W' stands for batch, channel, height, and width"
-                  "dimensions respectively. Convolution is applied on the 'H' and"
-                  "'W' dimensions.");
-    TVM_ATTR_FIELD(kernel_layout).set_default("IHW")
-        .describe("Dimension ordering of weight. Can be 'IHW', 'HWI', etc."
-                  "'I', 'H', 'W' stands for input_channel, height, and width"
-                  "dimensions respectively.");
-    TVM_ATTR_FIELD(out_dtype)
-        .set_default(NullValue<DataType>())
-        .describe("Output data type, set to explicit type under mixed precision setting");
-  }
-};
-
-
 /*! \brief Attributes used in winograd weight transformation operators */
 struct Conv2DWinogradWeightTransformAttrs :
     public tvm::AttrsNode<Conv2DWinogradWeightTransformAttrs> {
