+++ src/relay/op/nn/convolution.cc
@@ -1023,66 +1023,5 @@ Expr MakeDeformableConv2D(Expr data,
 TVM_REGISTER_GLOBAL("relay.op.nn._make.deformable_conv2d")
 .set_body_typed(MakeDeformableConv2D);
 
-// relay.nn.dilation2d
-TVM_REGISTER_NODE_TYPE(Dilation2DAttrs);
-
-template<typename T>
-Array<Array<Layout> > Dilation2DInferCorrectLayout(
-    const Attrs& attrs,
-    const Array<Layout>& new_in_layouts,
-    const Array<Layout>& old_in_layouts,
-    const Array<tvm::relay::Type> &old_in_types) {
-  const T* params = attrs.as<T>();
-
-  // We always make other operators to fit the layouts of convolution layers
-  // So this inference ignores all inputs
-  return Array<Array<Layout> >{{params->data_layout, params->kernel_layout},
-                               {params->data_layout}};
-}
-
-// Positional relay function to create dilation2d operator
-// used by frontend FFI.
-Expr MakeDilation2D(Expr data,
-                    Expr weight,
-                    Array<IndexExpr> strides,
-                    Array<IndexExpr> padding,
-                    Array<IndexExpr> dilations,
-                    std::string data_layout,
-                    std::string kernel_layout,
-                    DataType out_dtype) {
-  auto attrs = make_object<Dilation2DAttrs>();
-  attrs->strides = std::move(strides);
-  attrs->padding = std::move(padding);
-  attrs->dilations = std::move(dilations);
-  attrs->data_layout = std::move(data_layout);
-  attrs->kernel_layout = std::move(kernel_layout);
-  attrs->out_dtype = std::move(out_dtype);
-  static const Op& op = Op::Get("nn.dilation2d");
-  return CallNode::make(op, {data, weight}, Attrs(attrs), {});
-}
-
-
-TVM_REGISTER_GLOBAL("relay.op.nn._make.dilation2d")
-.set_body_typed(MakeDilation2D);
-
-
-RELAY_REGISTER_OP("nn.dilation2d")
-.describe(R"code(Computes grayscale dilation of 4D input and 3D filter.
-- **data**: This depends on the `layout` parameter. Input is 4D array of shape
-            (batch_size, in_channels, height, width) if `layout` is `NCHW`.
-- **weight**: (in_channels, height, width)
-- **out**:  This depends on the `layout` parameter. Output is 4D array of shape
-            (batch_size, channels, out_height, out_width) if `layout` is `NCHW`.
-)code" TVM_ADD_FILELINE)
-.set_attrs_type<Dilation2DAttrs>()
-.set_num_inputs(2)
-.add_argument("data", "Tensor", "The input tensor.")
-.add_argument("weight", "Tensor", "The weight tensor.")
-.set_support_level(2)
-.add_type_rel("Dilation2D", Dilation2DRel<Dilation2DAttrs>)
-.set_attr<FInferCorrectLayout>("FInferCorrectLayout",
-         Dilation2DInferCorrectLayout<Dilation2DAttrs>);
-
-
 }  // namespace relay
 }  // namespace tvm
